Add specs to test missing team_role or different owner system
